requirements.txt
anyio==4.12.1
APScheduler==3.6.3
asgiref==3.11.1
attrs==25.4.0
cachetools==4.2.2
certifi==2026.1.4
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
postgres==4.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
PyJWT==2.11.0
python-decouple==3.8
python-telegram-bot==22.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
